Refactor Complete
After digging through it I was able to drastically cut down the number
of times the damage calculation is run.
Before it would have been run
(# of moves on weapon * # of parts on monster)  * (1 + # of elements on
weapon) * 7
(The 7 comes from calculating the actual damage once for display and
then 6 times to find the min and max for each of the following (per
part, per move, total))
now it is
(# of moves on weapon * # of parts on monster)  * 4
(The 4 coming from calculating the actual damage once for display and
then 3 times to find the ranges for the same three as above)

Note: Although not mentioned above both of the above calculations are
actually off by a factor of three(add a * 3 to the end of each
equations). I currently see no reason why this should exist but it might
be a problem with how often the javascript is called to refresh and
recalculate even if nothing has changed. Further research would probably
be a good idea.

Realistically it can be cut down to just (# of moves on weapon * # of
parts on monster) damage calculations but that would require a larger
change than I want to make in this commit.