Html Refactor
Removed about 400 lines of html and replaced with an ng-repeat element.
This was done because manually adding to the html was error-prone
especially copy pasta style mistakes(See lines 161 & 165 of the old html
to see that the 'name' property was not properly changed, perfect
example of problems that can arise without realizing).

Also added in a small piece of infrastructure to help support
bowguns(specifically only calculating damage for shots the bowgun can
fire.